提取命名类型
